fix(Card): Export Card as styled component
