+++ packages/components/src/components/Card/index.tsx
@@ -1,10 +1,29 @@
-import React, { FC } from 'react';
-import { StyledCard } from './style';
+import styled from 'styled-components';
+import ThemeTools from '../../themes/CustomTheme/ThemeTools';
 
-const Card: FC<{
-    'data-testid'?: string;
-}> = props => {
-    return <StyledCard data-testid={props['data-testid']}>{props.children}</StyledCard>;
+type CardThemeType = {
+    background: string;
+    borderColor: string;
+    borderRadius: string;
+};
+
+const Card = styled.div`
+    position: relative;
+    background: ${({ theme }) => theme.Card.background};
+    border: 1px solid ${({ theme }) => theme.Card.borderColor};
+    border-radius: ${({ theme }) => theme.Card.borderRadius};
+    flex: 1 1 auto;
+`;
+
+const composeCardTheme = (themeTools: ThemeTools): CardThemeType => {
+    const { forms } = themeTools.themeSettings;
+
+    return {
+        background: forms.background,
+        borderColor: forms.borderColor,
+        borderRadius: themeTools.calculateRoundness(5),
+    };
 };
 
 export default Card;
+export { CardThemeType, composeCardTheme };
