+++ packages/components/src/themes/CustomTheme/genTheme.ts
@@ -38,7 +38,7 @@ import { composeTextualButton } from '../../components/TextualButton';
 import { composeProgressTheme } from '../../components/Progress/style';
 import { composeFormRowTheme } from '../../components/FormRow/style';
 import { composeAccordionTheme } from '../../components/Accordion/style';
-import { composeCardTheme } from '../../components/Card/style';
+import { composeCardTheme } from '../../components/Card';
 import { composePaginationTheme } from '../../components/Pagination/style';
 import { composeFileInputTheme } from '../../components/FileInput/style';
 import { composeColorDropTheme } from '../../components/ColorDrop';
